+++ ChangeLog.markdown
@@ -15,6 +15,7 @@ __N/A__
 * #204: Images resource should be able to retrieve only specific fields
 * #201: Added an index resource
 * #199: Allow /metadata for the meta data resource
+* #196: Added Amazon Simple Storage Service (S3) storage adapter
 * #192: Images resource should allow sorting
 * #190: Added support for filtering images on one or more image identifiers
 * #189: Prioritize original image mime types when doing content negotiation